Enhance Alpaca adapter with orderbook support and data type improvements
- Introduce orderbook subscription and unsubscription methods in the AlpacaWebSocketClient for real-time orderbook updates.
- Update AlpacaWsTrade, AlpacaWsQuote, and AlpacaWsBar structures to use Decimal for size and volume, accommodating fractional amounts in crypto trading.
- Modify AlpacaDataClient to handle orderbook messages and resubscribe logic.
- Enhance parsing functions to support new orderbook data types and improve overall data handling efficiency.
- Update example scripts to reflect changes in orderbook subscriptions and logging for better debugging.

+++ crates/adapters/alpaca/src/common/models.rs
@@ -43,9 +43,9 @@ pub struct AlpacaWsTrade {
     /// Trade price.
     #[serde(rename = "p")]
     pub price: Decimal,
-    /// Trade size (number of shares/units).
+    /// Trade size (number of shares/units). Decimal for crypto fractional amounts.
     #[serde(rename = "s")]
-    pub size: u64,
+    pub size: Decimal,
     /// Trade timestamp (RFC3339 format).
     #[serde(rename = "t")]
     pub timestamp: String,
@@ -74,18 +74,18 @@ pub struct AlpacaWsQuote {
     /// Ask price.
     #[serde(rename = "ap")]
     pub ask_price: Decimal,
-    /// Ask size.
+    /// Ask size. Decimal for crypto fractional amounts.
     #[serde(rename = "as")]
-    pub ask_size: u64,
+    pub ask_size: Decimal,
     /// Bid exchange code.
     #[serde(rename = "bx")]
     pub bid_exchange: Option<String>,
     /// Bid price.
     #[serde(rename = "bp")]
     pub bid_price: Decimal,
-    /// Bid size.
+    /// Bid size. Decimal for crypto fractional amounts.
     #[serde(rename = "bs")]
-    pub bid_size: u64,
+    pub bid_size: Decimal,
     /// Quote timestamp (RFC3339 format).
     #[serde(rename = "t")]
     pub timestamp: String,
@@ -120,9 +120,9 @@ pub struct AlpacaWsBar {
     /// Close price.
     #[serde(rename = "c")]
     pub close: Decimal,
-    /// Volume.
+    /// Volume. Decimal for crypto fractional amounts.
     #[serde(rename = "v")]
-    pub volume: u64,
+    pub volume: Decimal,
     /// Bar start timestamp (RFC3339 format).
     #[serde(rename = "t")]
     pub timestamp: String,
